VBE-library/requirements.txt
numpy==1.26.0
nltk==3.8.1
plotly==5.18.0
matplotlib==3.8.0
pandas==2.2.2
inquirer==2.7.0
scikit-learn==1.1
scipy==1.11.0
python-dotenv
psycopg2
requests==2.31.0
sqlalchemy==2.0.20